Split ProjectState.WithHostProject into into two methods
Currently, there is a single method for legal changes to a project's HostProject: ProjectState.WithHostProject. However, is is illegal for most of the properties on HostProject to be modified for a given project. For example, if the ProjectKey, FilePath or IntermediateOutputPath were to change, the HostProject would be inconsistent with maps in the ProjectSnapshotManager. To avoid potential bug farms like these, this change removes WithHostProject and replaces it with two methods WithConfiguration and WithRootNamespace, which are the only two values that we currently allow to be modified.

A particular challenge with this change is that the FallbackHostProject needs to be removed in order for tests to pass. This will happen in the next commit.