Make FallbackProjectManager to track fallback projects
Previously, fallback projects were tracked by being created with a FallbackHostProject. Eventually, this would be replaced with a standard HostProject if the project system started tracking the project and updated the project with a new configuration or root namespace.

Now that its not possible to update a project's HostProject directly, it's no longer viable to track whether a project is a fallback project via the HostProject. So, the FallbackProjectManager has been updated to track fallback projects itself as an ImmutableHashSet<ProjectKey>. Features that need to understand fallback projects can import an IFallbackProjectManager rather than digging into ProjectSnapshot.HostProject.