Seal HostProject and rework MiscFilesHostProject
HostProject was abstract to allow for the existence of FallbackHostProject and MiscFilesHostProject as descendants. Now that FallbackHostProject has been removed, it seems contrived for MiscFilesHostProject to exist, and it isn't needed by production code.

This change reworks MiscFilesHostProject into a static class: MiscFilesProject. This allows HostProject to be marked as sealed.