BoundAttributeDescriptor: Compute flags in builder
BoundAttributeDescriptor holds several property values in a "flags" field that is computed in the constructor. This change computes the flags BoundAttributeDescriptorBuilder instead. In addition, the JSON.NET and MessagePack serializers have been updated to write/read the flags field. Finally, "IsDirectiveAttribute" metadata has been removed, since that value is part of the flags.