RequiredAttributeDescriptor: Introduce flags and compute in builder
This change introduces a RequiredAttributeFlags enum to store the various values of RequiredAttributeDescriptor. In addition to CaseSensitive and IsDIrectiveAttribute, the NameComparison and ValueComparison values are packed into the flags. The flags are computed in RequiredAttributeDescriptorBuilder and passed to the RequiredAttributeDescriptor constructor. Finally, the JSON.NET and MessagePack serializers have been updated to write/read the flags field and "IsDirectiveAttribute" has been removed from metadata.