TagHelperDescriptor: Replace Runtime.Name metadata with property
Every TagHelperDescriptor was required to have a "Runtime.Name" value in its metadata. And, there was code to enforce that a tag helper  a missing "Runtime.Name" metadata value would have a default value added. Since all TagHelperDescriptors are required to have this data, this change just makes it a property on TagHelperDescriptor.