Resolve CSharpLanguageVersion from RazorParserOptions
It has always been necessary to call RazorProjectEngineBuilder.SetCSharpLanguageVersion(...) when configuring a RazorProjectEngine. However, this shouldn't be necessary, since the CSharpParseOptions are part of the RazorParserOptions.

This change removes SetCSharpLanguageVersion(...), which is a breaking change that will need to be fixed in the RazorSDK. After this change, RazorParserOptions.CSharpParseOptions.LanguageVersion is passed into the RazorCodeGenerationOptions.Builder. The logic to set various code-gen options based on the Razor and C# language versions has moved to RazorCodeGenerationOptions.GetDefaultFlags(...)

While it's certainly possible to set the CSharpParseOptions by calling the RazorProjectEngineBuilder.ConfigureParserOptions(...) method, I've added a helper RazorProjectEngineBuilder.SetCSharpParsOptions(...).