Rename RazorFileKind to RazorSourceCodeKind and clean up
This commit is huge, but it's almost completely mechanical. The purpose of this giant change is to rename the RazorFileKind enum to RazorSourceCodeKind and adjust names containing "FileKind" to "SourceCodeKind". This brings the enum in line with Roslyn's SourceCodeKind enum.

Buried in this change are tweaks to remove RazorFileKinds.GetFileKindFromFilePath(...). Callers have been replaced with TryGetSourceCodeKind(...).