Make TryGetProvisionalCompletionInfoAsync synchronous
This helper method is async, but the only reason for that is to call GetCodeDocumetnAsync(...). However, all of its callers already have a RazorCodeDocument in hand. So, this method can be made synchronous and avoid an async state machine by just passing a RazorCodeDocument to it rather than a DocumentContext. In addition, by making it synchronous, the method can be refactored into a bool-returning TryGet* method.