Clean up CompletionTriggerAndCommitCharacters
CompletionTriggerAndCommitCharacters recently converted to a singleton service from a static class, but much of it was still accessed statically. This change takes it the rest of the way.

Raw collections are no longer exposed from CompletionTriggerAndCommitCharacters. Instead, collections of characters are held in instance fields and methods are exposed that use them.  Now that the collections are private, I've switched them to HashSet<string>. It's highly unlikely that we were getting any value from FrozenSet<string>, since the values it contains are strings of the same same length. In addition, the AllCommitCharacters and AllTriggerCharacters properties have been converted to methods that return new arrays on each call. These are just called once in VS or Roslyn OOP, and it's better to guard against them becoming corrupted if something in the serialization modifies them in some way.