Don't store TagHelperDocumentContext in ItemCollection
Lift TagHelperDocumentContext out of RazorCodeDocument.Items and into a field on RazorCodeDocument.

Note: This change includes a nullability change to IRazorCodeDocument.GetTagHelperContext() in the legacy editor shim. It turns out that there's a single reference to this API in the legacy editor that already expects null, so this is safe.