DefaultDynamicDocumentContainer: Avoid holding onto DocumentSnapshot
DefaultDynamicDocumentContainer and the services it exposes cause a DocumentSnapshot to be held onto within the dynamic file info system. Because DocumentSnapshot references ProjectSnapshot, it keeps an entire project sanpshot and all other documents alive, include the RazorCodeDocuments within them. Even worse, because the TextLoader returned by DefaultDynamicDocumentContainer holds a DocumentSnapshot, the Roslyn workspace itself is holding Razor project snapshots alive!

This change updates DefaultDynamicDocumentContainer to hold a RazorCodeDocument instead of a DocumentSnapshot. The BackgroundDocumentGenerator was already forcing the RazorCodeDocument to be generated, so we can avoid hanging onto the whole project snapshot and everything within it. In order for this to work, all of the document-level services owned by DefaultDynamicDocumentContainer have been updated to use the same RazorCodeDocument rather than a DocumentSnapshot.

While working on this change, I also removed the IDocumentSnapshot.GetCSharpSyntaxTreeAsync(...) method. All of the callers can move over to GetOrParseCSharpSyntaxTree(...) or use RemoteDocumentSnapshot.GetCSharpSyntaxTreeAsync(...) (for co-hosting).