RazorDocumentServiceProvider: Use Optional<T> to avoid repeated requests
RazorDocumentServiceProvider uses null as the signal to request a service from the IDynamicDocumentContainer. However, it's perfectly legal for IDynamicDocumentContainer to return null, which means that it'll just get called again for the next service. By using Optional<T> we can avoid extra calls when the IDynamicDocumentContainer doesn't provide a service.