Remove ProjectSnapshotHandle
All of the infrastructure that's been added in support of co-hosting makes it much easier to call into Roslyn OOP to perform tag helper discovery. We really don't need the ProjectSnapshotHandle class. We can just pass the ProjectId and RazorConfiguration as parameters, and the root namespace was never actually used. In addition, we had old JSON serialization code for ProjectSnapshotHandle that was no longer used.