Cache SourceText in RazorDynamicFileInfo.TextLoader
Use an AsyncLazy to cache the SourceText returned by GeneratedDocumentTextLoader. This TextLoader is passed to Roslyn via RazorDynamicFileInfo and is used for the generated C# file that gets added in the workspace. However, the SourceText in the workspace is held in a weak reference. If the SourceText is collected, the workspace will call into our TextLoader again, causing Razor to compile again. By using an AsyncLazy, that TextLoader only compiles once. Note that this might ultimately increase committed memory because Razor doesn't have temporary storage to flush objects to MemoryMappedFiles, like Roslyn does.