Remove IRazorMappingService from CSharpTestLspServer
The CSharpTestLspServer provides create methods and parameters that can pass an IRazorMappingService. However, the only IRazorMappingService implementations that are ever passed in are dummy implementations that either throw or return empty arrays. All existing tests pass fine when the IRazorMappingService is removed, so it seems fine to leave it off until some future unit tests actually need it.