Improve IntermediateNode.FindDescendantNodes(...) extension method
Add an includeNode parameter to FindDescendentNodes<TNode>(...) that allows a caller to pass a delegate that will be called on matching nodes to determine whether they should be added to the results. This helps avoid List<T> growth when unnecessary.

Add a similar CollectDescendentNodes<TNode>(...) extension method that takes a List<TNode> to populate, rather than creating a new list.