Convert PooledHashSet<T> to [NonCopyable] struct with inline elements
The PooledHashSet<T> helper struct is used to acquire a HashSet<T> from an ObjectPool<T> only when items are added. However, it can avoid that acquisition for very small sets by storing the first two elements on the stack. This is based on the PooledArrayBuilder<T> implementation that's already in place.