Rewrite RuntimeNodeWriter.WriteHtmlContent to avoid allocations
`RuntimeNodeWriter.WriteHtmlContent` allocates a char array up front
to hold the HTML token content. Then, if the content is large enough, it
splits the content into chunks as it writes it out. This change rewrites
the method to avoid the upfront allocation and do the chunking entirely
with `PooledArrayBuilder<ReadOnlyMemory<char>>` instances. This uses
the new `CodeWriter.WriteStringLiteral(ReadOnlyMemory<char>)` extension
method to write out the content directly.