Change TagHelperDescriptorProviderContext.TargetSymbol to TargetAssembly
In tag helper discovery, there is an opportunity to provide an optional ISymbol to target during search. However, in all of the cases where a target symbol is provided, an IAssemblySymbol is passed, which is really the only thing that makes sense. This change switches the optional target ISymbol to an optional target IAssemblySymbol. This clarifies the intent a bit better.