Fix SyntaxNodeExtensions.WithAnnotations(...)
SyntaxNodeExtensions.WithAnnotations(...) shouldn't call CreateRed with a parent and position. Roslyn's version of this method doesn't do this, and it's really only intended to be called by the generated InternalSyntax nodes.

With this change in place, it's necessary to fix up the only place that calls WithAnnotations(...) (outside of Syntax.cml.Internal.Generated.cs) in LegacySyntaxNodeExtensions. With this in place, it's necessary to fix a code in SyntaxUtilities that creates red nodes unnecessarily. Instead, that code should just create green nodes to build a SyntaxList<RazorSyntaxNode> from. To make this easier, overloads have been added to SyntaxBuilderExtensions that operate on PooledArrayBuilder<GreenNode>.